+++ internal/tmdb/tv/type.go
@@ -7,15 +7,15 @@ type TV struct {
 	Status       string  `json:"status"`
 	Adult        bool    `json:"adult"`
 	Popularity   float32 `json:"popularity"`
-	Seasons      int64   `json:"number_of_seasons"`
-	Episodes     int64   `json:"number_of_episodes"`
+	Seasons      int32   `json:"number_of_seasons"`
+	Episodes     int32   `json:"number_of_episodes"`
 	BackdropPath string  `json:"backdrop_path"`
 	PosterPath   string  `json:"poster_path"`
 }
 
 type Season struct {
 	Id           int64     `json:"id"`
-	SeasonNumber int64     `json:"season_number"`
+	SeasonNumber int32     `json:"season_number"`
 	Name         string    `json:"name"`
 	Overview     string    `json:"overview"`
 	Episodes     []Episode `json:"episodes"`
@@ -24,7 +24,7 @@ type Season struct {
 type Episode struct {
 	Id            int64  `json:"id"`
 	AirDate       string `json:"air_date"`
-	SeasonNumber  int64  `json:"season_number"`
-	EpisodeNumber int64  `json:"episode_number"`
-	Runtime       int64  `json:"runtime"`
+	SeasonNumber  int32  `json:"season_number"`
+	EpisodeNumber int32  `json:"episode_number"`
+	Runtime       int32  `json:"runtime"`
 }
